+++ src/ui/event/tests.rs
@@ -6,19 +6,17 @@ use crate::WebEventKind;
 
 #[test] #[rustfmt::skip]
 fn sizes_of() {
-    /* event, kind */
+    /* event, kind, window */
 
     #[cfg(not(feature = "alloc"))] {
-    // assert_eq![40, size_of::<Event>()];          // 320 bits (old)
-    // assert_eq![44, size_of::<Event>()];          // 352 bits (+processed)
-    // assert_eq![64, size_of::<Event>()];          // 512 bits (+proc+count u64)
     assert_eq![56, size_of::<Event>()];             // 448 bits (+proc+count NonZeroU64)
     assert_eq![36, size_of::<EventKind>()];         // 288 bits
+    assert_eq![16, size_of::<EventWindow>()];       // 128 bits
     }
     #[cfg(feature = "alloc")] {
-    // assert_eq![48, size_of::<Event>()];          // 384 bits (old)
     assert_eq![56, size_of::<Event>()];             // 448 bits (+proc+count NonZeroU64)
     assert_eq![40, size_of::<EventKind>()];         // 320 bits
+    assert_eq![24, size_of::<EventWindow>()];       // 192 bits
     }
 
     /* key */
@@ -61,13 +59,6 @@ fn sizes_of() {
     /* timestamp */
 
     assert_eq![4, size_of::<EventTimestamp>()]; // 32
-
-    /* window */
-
-    #[cfg(not(feature = "alloc"))]
-    assert_eq![16, size_of::<EventWindow>()];   // 128
-    #[cfg(feature = "alloc")]
-    assert_eq![24, size_of::<EventWindow>()];   // 192
 }
 
 #[test]